Split remaining slow shell test suites
